New interface for type wrappers
